fix: Don't override evenr ID if undefined (#6316)
A serializer was created to convert string event identifier to event ID while saving the event or 
event relationship. However, event ID may have already been replaced sometimes and sometimes 
originalId is undefined, so in this case, valid event ID was being replaced with undefined. So, we 
check for null or undefined before replacing event ID