Fix incorrect treatment of promotional codes (Fixes: #8919)
If a promotional code is passed in, the code checks both
access code and discount code, and if one of those is not
valid, lists the code as invalid. Since access codes are checked
first, the later check for discount code will disable the
promotional code.

Use two separate booleans to check.